refactors import paths to use utils module for cache and dependencies
