+++ src/routes/sockets.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter
 from starlette.websockets import WebSocket, WebSocketDisconnect
 
 from src.connections import ConnectionManager, RedisConnectionManager
-from src.dependencies import Schedule, WebsocketConnectionManager, YahooCookies, YahooCrumb
+from utils.dependencies import Schedule, WebsocketConnectionManager, YahooCookies, YahooCrumb
 from src.models import MarketSector, SimpleQuote
 from src.security import validate_websocket
 from src.services import (
