+++ src/services/indices/fetchers/fetch_index.py
@@ -1,6 +1,6 @@
 from orjson import orjson
 
-from src.dependencies import fetch
+from utils.dependencies import fetch
 from src.models.index import Index, MarketIndex
