+++ src/services/movers/fetchers/movers_api.py
@@ -1,6 +1,6 @@
 from orjson import orjson
 
-from src.dependencies import fetch
+from utils.dependencies import fetch
 from src.models import MarketMover
