+++ src/services/movers/get_movers.py
@@ -1,4 +1,4 @@
-from src.cache import cache
+from utils.cache import cache
 from src.models.marketmover import MoverCount
 from src.services.movers.fetchers import fetch_movers, scrape_movers
