+++ src/services/news/get_news.py
@@ -1,8 +1,8 @@
 from fastapi import HTTPException
 from lxml import etree
 
-from src.cache import cache
-from src.dependencies import fetch
+from utils.cache import cache
+from utils.dependencies import fetch
 from src.models import News
