+++ src/services/similar/get_similar_quotes.py
@@ -1,4 +1,4 @@
-from src.cache import cache
+from utils.cache import cache
 from src.models import SimpleQuote
 from src.services.similar.fetchers import fetch_similar, scrape_similar_quotes
